// src/views/post-offer/post-offer.interface.ts
export interface PostJobProps {
    companyName: string;
    companySize: number;
    location: string;
    experience: string;
    title: string;
    salaryFrom: number;
    salaryTo: number;
    employmentType: string;
    jobDescription: string;
    aboutCompany: string;
    tech: string;
    webSite: string;
    address: string;
  }